restructured code. - moved `partial_dag_generator` into `generators/partial_dag_generator.hpp` - moved `partial_dag3_generator` into `generators/partial_dag3_generator.hpp`
